IndexGet should be a size_t not an int as it indexes the DataNodeList.
This may fix the crash I saw on frontera with 28 threads

+++ libgadget/treewalk.c
@@ -50,7 +50,7 @@ static struct data_index
 {
     int Task;
     int Index;
-    int IndexGet;
+    size_t IndexGet;
 } *DataIndexTable;
 
 /*Initialise global treewalk parameters*/
